agent/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
inflection==0.5.1
pydantic==2.6.4
pyyaml==6.0.1
retry==0.9.2
expiringdict==1.2.2
websocket-client==1.7.0
requests==2.31.0
sqlalchemy==2.0.28
websockets==12.0

# gpu related
torch==2.2.2
torchvision==0.17.2
transformers==4.39.3
sentencepiece==0.2.0
sacremoses==0.1.1

# dev
black